feat(elation): better helper funcs
Ref: #2498
